Correction and efficiency improvements - see extended description
When investigating a log warning about an automation already running, I realised I had two automations with the same name (the dangers of copy and paste!). Whilst fixing this I took the opportunity to also move the multiplier adjustment into a script. This means a lot less states need to be watched for the automation trigger and the calculation is only done once at midnight, not several times.