Quick fix for cycle 'will end' time.
Adding the possibility of 16 zones from 8 introduced a small bug which I think ONLY affects anyone who adds zones higher than 9 and then removes them. This is not an ideal solution but it works for me and at least one other user. As it only affects a small number of people and is non critical I plan to leave it at that for now. The change should have no effect for anyone not affected by the bug.